wip: langs + config

+++ src/test/generate/generate.cpp
@@ -26,6 +26,7 @@ int generate(int argc, const char* argv[]) {
         .value_type("SOURCE")
         .description("Generator executable or source")
         .store(opts.generator);
+
     parser
         .add('v', "var")
         .optional()
@@ -42,10 +43,10 @@ int generate(int argc, const char* argv[]) {
                 splitted.push_back("1");
             }
 
-            Range<int> range {
-                util::from_string<int>(splitted[1]),
-                util::from_string<int>(splitted[2]),
-                util::from_string<int>(splitted[3]),
+            Range<i64> range {
+                util::from_string<i64>(splitted[1]),
+                util::from_string<i64>(splitted[2]),
+                util::from_string<i64>(splitted[3]),
             };
 
             opts.vars[name] = range;
